refactor: 구현해둔 기능이 더 관련있는 컴포넌트로 이동
